Imported using git-ubuntu import.

Publish parent: 649ebb72f5d84aada08967eb07183431d7b4ff8b
Changelog parent: 390a59660157cb621a5bc48a989eef93941fa84a
Unapplied parent: f7f7f1d6da9519eb2f1e24a1e53988cf54240aa3

New changelog entries:
  * New upstream release:
    - debian/control: regenerate.
    - debian/symbols.wildcard: update for 2.25.
    - debian/patches/git-updates.diff: update from upstream stable branch.
    - debian/patches/alpha/local-string-functions.diff: rebased.
    - debian/patches/alpha/submitted-fegetenv-namespace.diff: upstreamed.
    - debian/patches/amd64/local-blacklist-for-Intel-TSX.diff: upstreamed.
    - debian/patches/hppa/cvs-atomic-machine.diff: upstreamed.
    - debian/patches/hurd-i386/tg-tls.diff: rebased.
    - debian/patches/hurd-i386/tg-tls-threadvar.diff: rebased.
    - debian/patches/hurd-i386/tg-hurdsig-fixes-2.diff: rebased.
    - debian/patches/hurd-i386/unsubmitted-NO_HIDDEN.diff: rebased.
    - debian/patches/hurd-i386/cvs-check-local-headers.diff: upstreamed.
    - debian/patches/hurd-i386/tg-faccessat.diff: rebased and reworked.
    - debian/patches/hurd-i386/tg-gsync-libc.diff: rebased.
    - debian/patches/hurd-i386/git-recvmsg.diff: upstreamed.
    - debian/patches/hurd-i386/cvs-adjtime-NULL.diff: upstreamed.
    - debian/patches/hurd-i386/cvs-mallocfork.diff: upstreamed.
    - debian/patches/hurd-i386/cvs-hurd_signal.h_c++.diff: upstreamed.
    - debian/patches/i386/cvs-cpuid-level2.diff: upstreamed.
    - debian/patches/kfreebsd/submitted-auxv.diff: rebased.
    - debian/patches/sparc/cvs-test-strncmp.diff: upstreamed.
    - debian/patches/any/local-rtlddir-cross.diff: rebased.
    - debian/patches/any/local-asserth-decls.diff: rebased.
    - debian/patches/any/submitted-longdouble.diff: dropped, obsolete.
    - debian/patches/any/local-disable-libnss-db.diff: rebased.
    - debian/patches/any/local-revert-bz13979.diff: rebased.
    - debian/patches/any/submitted-dlopen-noload.diff: upstreamed.
    - debian/patches/any/cvs-malloc-hardening.diff: rebased.
  * debian/testsuite-xfail-debian.mk: mark a few resolv related tests as
    XFAIL, as they fail due to local-dynamic-resolvconf.diff, which
    automatically reload a changed /etc/resolv.conf file. A better version
    of this patch has been merged in glibc 2.26, but is not easily
    backportable.
  * debian/testsuite-xfail-debian.mk: mark tst-strfrom and tst-strfrom-locale
    as XFAIL on mips*, they are due to corner case bugs in the FPU. Give more
    details about each issue.
  * debian/testsuite-xfail-debian.mk: mark test-double, test-idouble,
    test-ifloat and test-ildouble as XFAIL on s390x, they are due to a bug in
    GCC 6 (fixed in GCC 7) and not a regression from 2.24.

+++ sysdeps/ieee754/ldbl-128/s_logbl.c
@@ -26,15 +26,15 @@ static char rcsid[] = "$NetBSD: $";
 #include <math.h>
 #include <math_private.h>
 
-long double
-__logbl (long double x)
+_Float128
+__logbl (_Float128 x)
 {
   int64_t lx, hx, ex;
 
   GET_LDOUBLE_WORDS64 (hx, lx, x);
   hx &= 0x7fffffffffffffffLL;	/* high |x| */
   if ((hx | lx) == 0)
-    return -1.0 / fabs (x);
+    return -1.0 / fabsl (x);
   if (hx >= 0x7fff000000000000LL)
     return x * x;
   if ((ex = hx >> 48) == 0)	/* IEEE 754 logb */
@@ -48,7 +48,7 @@ __logbl (long double x)
 	ma = __builtin_clzll (hx);
       ex -= ma - 16;
     }
-  return (long double) (ex - 16383);
+  return (_Float128) (ex - 16383);
 }
 
 weak_alias (__logbl, logbl)
